グラフの描画を別コンポーネントにファイル分割
